all but two tests in c/test_c.py:test_complex_types() pass.
  1)   assert repr(complex(cast(p, -0j))) == '-0j'
  2)   assert complex(cast(p, '\x09')) == 9.0

I haven't looked at 2) yet but 1) is failing due to python's behavior: the answer should be -0j, it is returning 0j instead.

The issue is in complexobject.c: complex_new, at the end where it does this:
    if (cr_is_complex) {
        ci.real += cr.imag;
    }
    return complex_subtype_from_doubles(type, cr.real, ci.real);

cr.imag is appropriately "-0" still, but after this statement, ci.real is just 0. The resulting returned object is just 0j, not -0j.